streamlitMainpage/requirements.txt
asgiref==3.5.2
asn1crypto==0.24.0
astroid==2.9.0
altair==4.2.2
attrs==21.4.0
bcrypt==3.2.2
cached-property==1.5.2
certifi==2021.10.8
cffi==1.12.3
chardet==3.0.4
click==8.1.3
#conda==4.12.0
#conda-package-handling==1.3.11
cryptography==2.7
distro==1.7.0
docker==5.0.3
dockerpty==0.4.1
docopt==0.6.2
future==0.18.2
h11==0.13.0
idna==2.8
importlib-metadata==4.11.4
isort==5.9.3
Jinja2==3.1.2
jsonschema==3.2.0
lazy-object-proxy==1.6.0
libarchive-c==2.8
MarkupSafe==2.1.1
mccabe==0.6.0
numpy==1.20.3
paramiko==2.11.0
Pillow==8.2.0
pipreqs==0.4.13
platformdirs==2.4.0
pycosat==0.6.3
pycparser==2.19
pylint==2.12.2
PyNaCl==1.5.0
pyOpenSSL==19.0.0
pyrsistent==0.18.1
PySocks==1.7.0
python-dotenv==0.20.0
PyYAML==6.0
requests==2.22.0
#ruamel-yaml==0.15.46
six==1.12.0
streamlit==1.18.0
texttable==1.6.4
toml==0.10.2
#torch==1.5.1
#torchvision==0.6.1
tqdm==4.32.1
typed-ast==1.4.3
typing-extensions==4.1.1
urllib3==1.24.2
uvicorn==0.17.6
websocket-client==0.59.0
wrapt==1.13.3
yarg==0.1.9
zipp==3.8.0
